\name{tidy.didbjs}
\alias{tidy.didbjs}
\alias{tidy}
\title{Tidy a didbjs result}
\usage{
tidy(x, ...)
\method{tidy}{didbjs}(x, ...)
}
\arguments{
\item{x}{A \code{didbjs} result.}
\item{...}{Unused.}
}
\description{
Return the coefficient table from a \code{didbjs} result.
}
\value{
A data frame of estimates.
}
